+++ modules/@apostrophecms/characterCount/ui/apos/components/TiptapCharacterCount.vue
@@ -36,115 +36,56 @@
 </template>
 
 <script>
+import characterCountMixin from '../mixins/characterCountMixin';
 
 export default {
   name: 'TiptapCharacterCount',
+  mixins: [characterCountMixin],
   props: {
-    options: {
-      type: Object,
-      required: true
-    },
     tool: {
-      type: Object,
-      required: true
-    },
-    editor: {
       type: Object,
       required: true
     }
   },
   data() {
     return {
-      generation: 1,
       active: false,
-      triggerValidation: false,
-      docFields: {
-        data: {}
-      },
-      formModifiers: ['small', 'margin-micro'],
-      totalCharactersCount: 0,
-      totalWordsCount: 0,
       highlightedCharacters: 0,
       highlightedWords: 0
     };
   },
   computed: {
-    moduleOptions() {
-      return apos.modules[apos.area.widgetManagers['@apostrophecms/rich-text']].aposCharCountConfig;
-    },
-    widgetOptions() {
-      if (this.options?.charCountConfig) {
-        return this.options.charCountConfig;
-      }
-      return {};
-    },
     buttonActive() {
       return this.active;
     },
     lastSelectionTime() {
       return this.editor.view.lastSelectionTime;
-    },
-    editorLimitText() {
-      if (this.moduleOptions?.limit || this.widgetOptions?.limit) {
-        const limit = this.widgetOptions.limit ? this.widgetOptions.limit : this.moduleOptions.limit;
-        return `/${limit}`;
-      }
-      return '';
-    },
-    hasSelection() {
-      const { state } = this.editor;
-      const { selection } = this.editor.state;
-      const { from, to } = selection;
-      const text = state.doc.textBetween(from, to, '');
-      return text !== '';
     }
   },
   mounted() {
-    this.calculateTotalCharacters();
-    this.calculateTotalWords();
     this.calculateHighlightedWords();
     this.calculateHighlightedCharacters();
   },
   watch: {
     active(newVal) {
       if (newVal) {
-        this.calculateTotalCharacters();
-        this.calculateTotalWords();
         this.calculateHighlightedWords();
         this.calculateHighlightedCharacters();
-        window.addEventListener('keydown', this.keyboardHandler);
-      } else {
-        window.removeEventListener('keydown', this.keyboardHandler);
       }
     },
   },
   methods: {
+    close() {
+      if (this.active) {
+        this.$emit('done');
+        this.$emit('cancel');
+      }
+    },
     takeAction() {
       this.active = !this.active;
       if (this.active) {
         this.populateFields();
       };
-    },
-    close() {
-      if (this.active) {
-        this.active = false;
-        this.editor.chain().focus();
-      }
-    },
-    keyboardHandler(e) {
-      if (e.keyCode === 27 || e.keyCode === 13) {
-        this.close();
-        e.preventDefault();
-      }
-    },
-    async populateFields() {
-      this.generation++;
-    },
-    calculateTotalCharacters() {
-      this.totalCharactersCount = this.editor.commands.getTotalCharactersCount();
-    },
-    calculateTotalWords() {
-      this.totalWordsCount = this.editor.commands.getTotalWordsCount();
     },
     calculateHighlightedCharacters() {
       this.highlightedCharacters = this.hasSelection ? this.editor.commands.getHighlightedStats('characters') : 0;
@@ -157,46 +98,11 @@ export default {
 </script>
 
 <style lang="scss" scoped>
-.apos-cc-control {
-  position: relative;
-  display: inline-block;
-}
-
-.apos-cc-control__dialog {
-  z-index: $z-index-modal;
-  position: absolute;
-  top: calc(100% + 5px);
-  left: -15px;
-  width: 250px;
-  opacity: 1;
-  pointer-events: none;
-  border: 1px solid var(--a-base-3);
-  border-radius: 3px;
-  background-color: white;
-}
-
-.character-count {
-  padding: 10px;
-  font-size: 12px;
-  line-height: 1.5;
-}
+@import '../scss/characterCountStyles';
 
 .apos-cc-control__dialog.apos-is-triggered.apos-has-selection {
   opacity: 1;
   pointer-events: auto;
 }
 
-.apos-is-active {
-  background-color: var(--a-base-7);
-}
-
-.apos-cc-control__footer {
-  display: flex;
-  justify-content: flex-end;
-  margin: 10px 10px 10px 0;
-}
-
-.apos-cc-control__footer .apos-button__wrapper {
-  margin-left: 7.5px;
-}
 </style>
