+++ modules/@apostrophecms/characterCount/ui/apos/scss/_characterCountStyles.scss
@@ -1,0 +1,38 @@
+.apos-cc-control {
+  position: relative;
+  display: inline-block;
+}
+
+.apos-cc-control__dialog {
+  z-index: $z-index-modal;
+  position: absolute;
+  top: calc(100% + 5px);
+  left: -15px;
+  width: 250px;
+  opacity: 1;
+  pointer-events: none;
+  border: 1px solid var(--a-base-3);
+  border-radius: 3px;
+  background-color: white;
+}
+
+.character-count {
+  padding: 10px;
+  font-size: 12px;
+  line-height: 1.5;
+}
+
+.apos-is-active {
+  background-color: var(--a-base-7);
+}
+
+.apos-cc-control__footer {
+  display: flex;
+  justify-content: flex-end;
+  margin: 10px 10px 10px 0;
+}
+
+.apos-cc-control__footer .apos-button__wrapper {
+  margin-left: 7.5px;
+}
+
